requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-filter==21.1
django-jazzmin==2.4.9
mysqlclient==2.1.0
Pillow==9.0.1
sqlparse==0.4.2
